fix: :recycle: moved API_URL in the apiClient file for easy access
